#76 Rename module (constants → consts)
