+++ arxiv_post/cli.py
@@ -10,7 +10,7 @@ from fire import Fire
 # submodules
 from .apps import slack
 from .arxiv import search
-from .constants import (
+from .consts import (
     CATEGORIES,
     KEYWORDS,
     START_DATE,
